Code cleanup (#103)
